--- temas/anotaciones/C#/CodeJSON/CodeJSON.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-  <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-</ItemGroup>
-
-
-</Project>
