--- temas/aspectos/C#/Code/ArchivoPolicial.cs
@@ -1,4 +1,0 @@
-﻿public interface ArchivoPolicial{
-    public void mostrarDetalles();
-    public void solicitarRevision();
-}
