+++ temas/aspectos/csharp/Code/ArchivoPolicial.cs
@@ -1,0 +1,4 @@
+﻿public interface ArchivoPolicial{
+    public void mostrarDetalles();
+    public void solicitarRevision();
+}
