+++ temas/aspectos/csharp/Code/Juicio.cs
