+++ temas/aspectos/csharp/Code/RegistradorAcciones.cs
