+++ temas/aspectos/csharp/Code/RegistroAspecto.cs
