+++ temas/inyeccion/csharp/Code/Code.csproj
@@ -1,0 +1,25 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <!-- Referencia para usar Xunit -->
+    <PackageReference Include="xunit" Version="2.8.0" />
+    <!-- Referencia para permitir ejecución en Visual Studio y otros runners -->
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.8.0" />
+    
+    <!-- Referencia a Newtonsoft.Json -->
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!-- Agrega el paquete para Microsoft.Extensions.DependencyInjection -->
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" /> <!-- Ajusta según tu versión de .NET -->
+  </ItemGroup>
+
+</Project>
