+++ temas/inyeccion/csharp/Code/Habitacion.cs
