+++ temas/inyeccion/csharp/Code/Salon.cs
